🧹 [Code Health] Centralize YAML parsing tools into utils.js (#23)
* Refactor: Centralize YAML parsing and data structure functions in scripts/utils.js

This commit moves the duplicated functions `parseMetadataYaml`, `parseScalar`, and `setNested`
from `scripts/validate-metadata.js` and `scripts/generate-manifest.js` to a central utility file
`scripts/utils.js`. It also updates tests and script imports to reflect this change, improving
code maintainability and removing duplicated logic.


* Merge master and resolve conflicts


---------
