+++ scripts/validate-metadata.test.js
@@ -361,3 +361,11 @@ test('listMetadataFiles', async (t) => {
     assert.deepStrictEqual(result, []);
   });
 });
+
+test('listMetadataFiles', async (t) => {
+  await t.test('returns empty array for non-existent directory', () => {
+    const nonExistentPath = '/path/that/does/not/exist/for/sure/12345';
+    const result = listMetadataFiles(nonExistentPath);
+    assert.deepStrictEqual(result, []);
+  });
+});
